+++ src/main/kotlin/klib/extensions/JsonArray.kt
@@ -1,6 +1,6 @@
 package klib.extensions
 
-import net.jemzart.jsonkraken.values.JsonArray
+import net.jemzart.jsonkraken.JsonArray
 
 inline fun <reified T> JsonArray.toListOfType(): List<T> {
     val typeOf: MutableList<T> = ArrayList()