+++ src/main/kotlin/net/jemzart/jsonkraken/constants/Escapable.kt
@@ -1,8 +1,0 @@
-package net.jemzart.jsonkraken.constants
-
-internal class Escapable {
-    companion object {
-        val monoChars = arrayOf('\"', '\\', '/', 'b', 'f', 'n', 'r', 't')
-        val whiteSpaceChars = arrayOf('\n', '\t', '\r')
-    }
-}