+++ src/main/kotlin/net/jemzart/jsonkraken/exceptions/NonCompliantStringException.kt
@@ -1,9 +1,0 @@
-package net.jemzart.jsonkraken.exceptions
-
-/**
- * An exception used when the String added to a JsonValue is not compliant with the JSON specification.
- *
- * @property value invalid string which triggered the exception.
- */
-class NonCompliantStringException(val value: String, message: String) :
-    Exception("String is not valid for JSON specification.\n$message")